Add option defaults to calcdeps()
If you use calcdeps in a node project, not using the CLI, e.g.

    calcdeps = require('calcdeps').calcdeps({ input: '' }, callback)

It will throw and error 'Cannot call method 'forEach' of undefined'. 
dependencies.forEach causes that when no dep option is set. 
This fixes that issue by providing a default array.
Beyond this it might be good to move all defaults from the bin to the lib version?